Moved `foreach.h` to `Dev/`; removed alignof/as from const_kvp
